optimized the reset code and add a custom EphemeryException
